Added parsing and building of type definitions.  Fixed up polymorphic instantiation code for function and constructor definitions.  Added explicit bitcasts and implicit upcasts.
TODO: Fix the generation of basic blocks when compiling an if expression so that the phi-node has a link back to every predecessor basic block (basic-block that branches to it).  Also, maybe give if expressions the same syntax as in C-family languages.  Lispy cond-style if's seem nice, but when pattern-matching is available I don't end up feeling the need to use them very much.