More progress towards compiling memes.dca, this time largely from fixing infinite-loop bugs in the type inference engine.
Honestly, I hadn't thought it would, but this type-inference engine repeats a lot of work and could really do with a full or partial rewrite.  Ironically, the version given in my paper is substantially simpler because it passes the constraint set as an immutable parameter to the inference function, so a step can be performed by composing a substitution onto the result of a recursive call to the inference procedure.