Commented out a lot of stuff from memes.dca to produce a working example of where the compiler is now.
Also fixed an infinite-recursion bug dealing with compiling the bodies of recursive functions.  What I really wanted as an "initialized" memoized function, which calculates the final result (the LLVMFunction object), stores it, and then runs a one-time initialization function on it once it's already memoized (the code to actually build the function body, which could recursively retrieve the memoized result).